Added usID as a parameter, so you can only delete it if the usID matches UserID AND if lID matches ListID

+++ deleteList.sql
@@ -1,12 +1,14 @@
 DELIMITER //
-DROP PROCEDURE IF EXISTS "deleteList" //
+DROP PROCEDURE deleteList //
 
 CREATE PROCEDURE deleteList(
-	in lID int
+	in lID int,
+	usID varchar(64)
 )
 
 BEGIN
 	DELETE FROM ToDoLists
-		WHERE ListID = lID;
+		WHERE (ListID = lID
+		AND UserID = usID);
 END//
 DELIMITER ;
